pyproject.toml
[project]
name = "braindump"
version = "0.1.0"
description = "Synthesize agent context from code review history"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Douwe Maan" }
]
keywords = ["claude", "ai", "code-review", "documentation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "typer>=0.12",
    "rich>=13",
    "sqlite-utils>=3.36",
    "lancedb>=0.27.0",
    "pyarrow>=23.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "logfire>=2.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
all = ["braindump[anthropic,openai]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pyright>=1.1",
    "ruff>=0.8",
]

[project.scripts]
braindump = "braindump.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/braindump"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
